pyproject.toml
[project]
name = "quizme-qdrant-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.134.0",
    "python-dotenv>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-qdrant>=1.1.0",
    "langchain-huggingface>=1.2.0",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3"]
